requirements.txt
# Deepgram only work for Python 3.10
python-dotenv==1.0.1
assemblyai
deepgram-sdk
groq